+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 
 ## Version 2025.3
 
+### Next release
+
+* 🔧 Debug log: as a safeguard introduce a limit for the number of messages.
+
 ### 2025.3.3 - 2025-11-20 🧪
 
 * 🔧 Debug log: change to always on, keep messages on app restarts and up to 3 days.